requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
asgiref==3.3.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cloudinary==1.17.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.6
django-cloudinary-storage==0.2.3
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-seed==0.2.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
Faker==5.6.5
filelock==3.0.12
future==0.18.2
html5lib==1.1
idna==2.10
mock==4.0.3
MouseInfo==0.1.3
mysql-connector-python==8.0.22
mysqlclient==2.0.1
phonenumbers==8.12.15
Pillow==8.0.1
psycopg2==2.8.6
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyPDF2==1.26.0
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-bidi==0.4.2
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.4
pywhatkit==3.8
reportlab==3.5.59
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.26.2
virtualenv==20.2.2
webencodings==0.5.1
whitenoise==5.2.0
wikipedia==1.4.0
xhtml2pdf==0.2.5
